+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Add the ability to set roles and affiliations via the MUC occupant modal
 - Fix `isOnlyEmojis is not a function` when using only `@converse/headless`
 - Fix `autojoin` checkbox state in MUC bookmark form
+- Remove call to `api.confirm` in `@converse/headless`
 
 ## 10.1.2 (2023-02-17)